Cleanup LUDCL cache invalidation and restoration logic
This cleans up the invalidation and restoration of cached
'latest user defined classloader' references when passing
between user and platform code during object deserialization.

We also take this opportunity to remove the unnecessary
`isClassCachingEnabled` instance variable and replace its
usage with the equivalent `classByNameCache != null`.

It is known that this fixes #20916, and it may also fix #9912
but that is as yet unvalidated.

Signed-off-by: Chris Dennis <chris.w.dennis@ibm.com>
(cherry picked from commit 7ce8262620a26dac48a8548b70f42b160ab71415)